Doing some error catching

+++ scripts/remove/nextcloud.sh
@@ -3,10 +3,15 @@
 echo -n -e "Please enter mysql root password so that nextcloud database and user can be dropped.\n"
 read -r -s -p "Password: " 'password'
 echo
-rm -rf /srv/nextcloud
-rm /etc/nginx/apps/nextcloud.conf
-systemctl reload nginx
 host=$(mysql -u root --password="$password" --execute="select host from mysql.user where user = 'nextcloud';" | grep -E "localhost|127.0.0.1")
 mysql --user="root" --password="$password" --execute="DROP DATABASE nextcloud;"
 mysql --user="root" --password="$password" --execute="DROP USER nextcloud@$host;"
+if [[ $? != "0" ]]; then 
+    echo "MySQL Drop failed. Please try again or investigate"
+    exit 1
+fi
+rm -rf /srv/nextcloud
+rm /etc/nginx/apps/nextcloud.conf
+systemctl reload nginx
+
 rm /install/.nextcloud.lock
